badger: do not allow force kernel extraction to change
Prevent unbootable machines by not allowing a kernel snap to be
installed that has force-kernel-extraction set when it was previously
unset and vice versa. It is not possible to enforce this in any
snap hook, as the bootloader is setup prior to any snap hooks
being run (which run on the new kernel after rebooting). Instead
of inventing some hook that could do this, just hard-code it to
be impossible to alter force-kernel-extraction.

The correct way to have done this would have been to use a different
name for the the encrypted kernel vs. the non-encrypted one, and then
to have the model key on that name. Then snapd would have already
kept us from bricking by only allowing the same named kernel snap.
Because we have used 'bar-kernel' for both encrypted and non-encrypted
snaps (and therefore extracted vs non-extracted ones), we need this
change to prevent accidental side-loading of the wrong kernel snap.
The consequence of such a mistake is a non-bootable machine.